Revert "fix(SETUPRACE-1): atomic fingerprint claim for setup-pause notifications (close D-0053)"
This reverts commit 31f6e87006ee61893c51b6bc5cc24876893ba108.
