+++ apps/core/src/adapters/storage/postgres/schema/canonical-ops-repo.postgres.ts
@@ -330,13 +330,6 @@ export class PostgresRuntimeRepositoryBundle
     return this.jobs.markJobSetupNotified(id, expectedFingerprint);
   }
 
-  async clearJobSetupNotified(
-    id: string,
-    expectedFingerprint: string,
-  ): Promise<boolean> {
-    return this.jobs.clearJobSetupNotified(id, expectedFingerprint);
-  }
-
   async deleteJob(id: string): Promise<void> {
     await this.jobs.deleteJob(id);
   }
