+++ apps/core/test/unit/application/job-management-create.test.ts
@@ -53,7 +53,6 @@ describe('job creation', () => {
           return true;
         },
       ),
-      clearJobSetupNotified: vi.fn(async () => true),
     };
     const setupRequiredNotifications: JobSetupRequiredNotificationPort = {
       notify: (input: JobSetupRequiredNotificationInput) => {
