+++ apps/core/test/unit/jobs/execution-finalization.test.ts
@@ -46,7 +46,6 @@ function makeDeps(): {
     opsRepository: {
       updateJob,
       markJobSetupNotified: vi.fn(async () => true),
-      clearJobSetupNotified: vi.fn(async () => true),
     },
     sendMessage,
     onSchedulerChanged: vi.fn(),
