+++ apps/core/test/unit/jobs/execution.test.ts
@@ -152,7 +152,6 @@ function makeOpsRepository(job: Job) {
     createJobRun: vi.fn(async () => true),
     updateJob: vi.fn(async () => undefined),
     markJobSetupNotified: vi.fn(async () => true),
-    clearJobSetupNotified: vi.fn(async () => true),
     completeJobRun: vi.fn(async () => undefined),
     finalizeJobRunLease: vi.fn(async (input) => {
       await repo.completeJobRun(
